cmd/destroy_test.go
package cmd

import (
	"testing"

	"github.com/winiceo/genv/internal/db"
	"github.com/winiceo/genv/pkg/container"
	"github.com/winiceo/genv/test_pkg"
)

func TestDestroy(got *testing.T) {
	t := test_pkg.NewT(got)

	cnt := container.Metadata{
		ID:        "foocnt",
		ImageID:   "fooimg",
		BaseName:  "fooenv",
		BaseImage: "scratch",
		Shell:     "/foo/sh",
		Mount: container.Mount{
			Destination: "/foo/mnt",
		},
	}

	s := &memStore{
		env: db.Environment{
			Status:    db.StatusReady,
			Container: cnt,
		},
	}

	ctl := newMockCtl(&cnt)

	cmd := newDestroyCmd(ctl, s)

	// Hijacking here swallows the command output so that it doesn't clutter
	// the output of `go test -v ./...`.
	outch, errch := test_pkg.HijackStdout(func() {
		cmd.Run(cmd, []string{})
	})

	select {
	case err := <-errch:
		t.Fatal("hijacking output", nil, err)
	case <-outch:
	}

	if nil != ctl.current {
		t.Fatal("backing container", nil, ctl.current)
	}

	if db.StatusOff != s.env.Status {
		t.Fatal("status", db.StatusOff, s.env.Status)
	}
}
